refactor(dal): migrate ids to long and add hybrid Dapper/SqlKata support

+++ TradeHub.BLL/Common/IIdentityService.cs
@@ -2,8 +2,6 @@
 {
     public interface IIdentityService
     {
-        int? UserId { get; }
-        //bool IsAuthenticated { get; }
-        // Sau này muốn lấy Email hay Role thì cứ thêm vào đây
+        long? UserId { get; }
     }
 }
